Google/Facebook login token refreshing http://nordland.oriconsys.com/issues/451
Consider the scenario: When the user logged in and stay on the same screen until the token is expired, the user may not use Cognito & Zuzu API with the expired token.

Use a timer to avoid the issue by the following handling:

1) Every time login is completed or session is resumed:
a. Cancel previous timer)
b. Set a timer to refresh the token
Trigger time is token's current expiry time plus 1 seconds

2) Cancel the timer when the user logs out (not in current UI design)

Note: When the login & resume process failed for some reason, the FB/GOOGLE token would be cleared (Log out)
The user will be prompted an error message, and he need to logged in again by the UI