雷達試用產品機制實作  http://nordland.oriconsys.com/issues/480
Do not display free trial if it's already used.

We only use UserDefault to check if the user has activated the free trial when deciding whether the product should be displayed.

The user may reinstall the App, and clear the UserDefault, but we'll check the status with the backend server before activating the free trial